feat: get distances to next pipe
Add a method in "Pipes" class to get their distance (X axis and Y axis) relative to the bird, in pixels. The window dimentions are 288x512.

These will be input to the neural network. It doesn't need any other information to be able to learn how to play.

Yet, they are just being printed in real time on the terminal.