feat: neural network synthesis
This is a neural network working with the 3D custom activation function, in which the weight and input are parameters to a function that outputs the next neuron value.

I understand it's bad written and hardcoded. The goal is only to test the capability of such a network to learn, and it's going to be tested only here. I don't have any experience building neural networks from scratch, but only know the principles and tried to implement them here.

Details in pages 52 to 55 of my field notebook.